+++ hk_fe_3/src/components/Header.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { useState } from "react";
 import {
   HeaderButtonDiv,
   HeaderDiv,
@@ -8,11 +9,13 @@ import {
 import { HeaderLogin } from "./../styles/Header/HeaderStyles";
 
 const Header = () => {
+  const [isSign, setIsSign] = useState(true);
+
   return (
     <HeaderDiv>
       <HeaderTitle>루튼 토마토</HeaderTitle>
       <HeaderButtonDiv>
-        <HeaderLogin>Sign In</HeaderLogin>
+        <HeaderLogin>{isSign ? "Sign Out" : "Sign In"}</HeaderLogin>
         <HeaderSignUp>Sign Up</HeaderSignUp>
       </HeaderButtonDiv>
     </HeaderDiv>
